refactor: Lottos로 함수 이동
LottoResult에 있던 Companion object의 함수를 Lottos 클래스로 이동
