+++ src/main/kotlin/lotto/domain/LottoResult.kt
@@ -1,6 +1,6 @@
 package lotto.domain
 
-class LottoResult private constructor(private val result: Map<Prize, Int>) : Map<Prize, Int> by result {
+class LottoResult(private val result: Map<Prize, Int>) : Map<Prize, Int> by result {
     private val totalPrizeMoney: Long
         get() {
             return result.map {
@@ -14,19 +14,5 @@ class LottoResult private constructor(private val result: Map<Prize, Int>) : Map
 
     fun getRateOfReturn(money: Int): Float {
         return totalPrizeMoney.toFloat() / money
-    }
-
-    companion object {
-        fun makeLottoResult(
-            winningLotto: Lotto,
-            bonusLottoNumber: LottoNumber,
-            lottos: Lottos,
-        ): LottoResult {
-            val result = lottos.lottos.groupingBy { lotto ->
-                val matchCount = lotto.countMatch(winningLotto)
-                Prize.of(matchCount = matchCount, bonusLottoNumber = bonusLottoNumber, lotto = lotto)
-            }.eachCount()
-            return LottoResult(result)
-        }
     }
 }
