+++ src/main/kotlin/lotto/domain/Lottos.kt
@@ -3,6 +3,17 @@ package lotto.domain
 class Lottos(lottos: List<Lotto>) {
     val lottos: List<Lotto> = lottos.toList()
 
+    fun getResult(
+        winningLotto: Lotto,
+        bonusLottoNumber: LottoNumber,
+    ): LottoResult {
+        val result = lottos.groupingBy { lotto ->
+            val matchCount = lotto.countMatch(winningLotto)
+            Prize.of(matchCount = matchCount, bonusLottoNumber = bonusLottoNumber, lotto = lotto)
+        }.eachCount()
+        return LottoResult(result)
+    }
+
     companion object {
         private const val LOTTO_PRICE = 1000
